requirements.txt
# HTTP请求和网络通信
requests>=2.31.0
urllib3>=1.26.16
aiohttp>=3.8.5

# 文件处理和图像处理
Pillow>=10.0.0
opencv-python>=4.8.0
python-magic>=0.4.27

# 任务调度和异步处理
APScheduler>=3.10.4
asyncio>=3.4.3

# 数据处理和存储
pydantic>=2.4.0
PyYAML>=6.0.1
cryptography>=41.0.4

# 日志和监控
loguru>=0.7.2

# 用户代理和反检测
fake-useragent>=1.4.0
selenium-stealth>=1.0.6

# 配置管理
python-dotenv>=1.0.0

# 开发和测试工具（可选）
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.7.0
flake8>=6.0.0 